KG/requirements.txt
rdflib
python-chess
datasets
pandas